PODS-2428: Move addTrusteeHeader into subclasses
